compile errors fix for tests

+++ x/claim/types/acitivity.go
@@ -3,7 +3,7 @@ package types
 func ActionToNames(actions []Action) []string {
 	names := []string{}
 	for _, action := range actions {
-		names = append(names, Action_name[action])
+		names = append(names, Action_name[int32(action)])
 	}
 	return names
 }
